+++ modules/portal-cache/portal-cache-test/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Portal Cache Test
+Bundle-SymbolicName: com.liferay.portal.cache.test
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.cache.test
+Include-Resource: classes
