+++ modules/portal-cache/portal-cache-test/src/com/liferay/portal/cache/test/TestPortalCache.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.cache.test;
 
-import com.liferay.portal.kernel.cache.AbstractPortalCache;
+import com.liferay.portal.cache.AbstractPortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
 
 import java.io.Serializable;
