+++ modules/settings.gradle
@@ -89,6 +89,7 @@ FileTree fileTree = fileTree(rootDir) {
 		include "portal-cache/portal-cache-ehcache-cluster/build.gradle"
 		include "portal-cache/portal-cache-ehcache-provider/build.gradle"
 		include "portal-cache/portal-cache-ehcache/build.gradle"
+		include "portal-cache/portal-cache-test/build.gradle"
 		include "portal-cache/portal-cache/build.gradle"
 		include "portal-security/*/build.gradle"
 		include "portal-store/*/build.gradle"
